requirements.txt
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.5.2
apprise==1.9.3
asgi-lifespan==2.1.0
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.29.0
attrs==25.3.0
backports.zoneinfo==0.2.1
boto3==1.37.38
botocore==1.37.38
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
croniter==5.0.1
cryptography==45.0.3
dateparser==1.2.1
dnspython==2.6.1
docker==7.1.0
durationpy==0.10
email-validator==2.2.0
exceptiongroup==1.3.0
filelock==3.16.1
fsspec==2025.3.0
google-auth==2.40.2
graphviz==0.20.3
greenlet==3.1.1
griffe==1.4.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.10.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==31.0.0
mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
orjson==3.10.15
packaging==24.2
pathspec==0.12.1
pendulum==2.1.2
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
prefect==2.20.18
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.1
PyJWT==2.9.0
pyopenssl==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
pytzdata==2020.1
PyYAML==6.0.2
readchar==4.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.20.1
rsa==4.9.1
ruamel.yaml==0.18.12
ruamel.yaml.clib==0.2.8
s3transfer==0.11.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowflake-connector-python==3.14.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.35
text-unidecode==1.3
toml==0.10.2
tomlkit==0.13.2
typer==0.15.4
typing-extensions==4.13.2
tzlocal==5.2
ujson==5.10.0
urllib3==1.26.20
uvicorn==0.33.0
websocket-client==1.8.0
websockets==13.1
zipp==3.20.2
